+++ src/containerapp/HISTORY.rst
@@ -6,6 +6,11 @@ upcoming
 ++++++
 * 'az containerapp up/create/update': Wait longer time for logstream of Cloud Build to make sure the container start
 
+0.3.51
+++++++
+* 'az containerapp env java-component config-server-for-spring': Rename spring-cloud-config to config-server-for-spring
+* 'az containerapp env java-component eureka-server-for-spring': Rename spring-cloud-eureka to eureka-server-for-spring
+
 0.3.50
 ++++++
 * 'az containerapp env telemetry data-dog show': Support show environment data dog configuration